+++ test/helpers/_helper.js
@@ -1,6 +1,6 @@
 import { base } from '../../src'
 import _get from 'lodash/get'
-import { BadRequestError } from '../../src/errors'
+import { BadRequestError, TransactionError } from '../../src/errors'
 
 export class Helper {
   /**
@@ -43,7 +43,7 @@ export class Helper {
 
     try {
       const { data } = await this.api.postTxEnvelope(envelope)
-      return data
+      return data.data.attributes
     } catch (e) {
       if (e instanceof BadRequestError) {
         const errorString = this.tryGetTxErrorString(e)
@@ -52,12 +52,15 @@ export class Helper {
         }
         throw e
       }
+      if (e instanceof TransactionError) {
+        throw new Error(e.errorResults())
+      }
       throw e
     }
   }
 
   tryGetTxErrorString (errorObject) {
-    let resultCodes = _get(errorObject, 'meta.extras.resultCodes')
+    let resultCodes = _get(errorObject, 'meta.resultCodes')
 
     if (resultCodes) {
       const txCode = resultCodes.transaction